+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## unreleased
   - New progressbar, using `tqdm`
   - `plot_families`:
+    - show average trace
     - pan/zoom traces using arrows
     - show/hide theoretical arrivals using 'a' (command line option removed)
     - show event origin time as y-label